analytics: fix people analytics monitor bug
If a trigger is set for a monitor and a Redis message for frame
metadata has never come in for that monitor, and apply_triggers is
called, there would be a KeyError for the channel.

Fix is to default to an empty event list if a trigger's channel
has received no events.

Change-Id: I37d5468a54c5c52ce7e091acdfcadda58a31686f