Following changes are there in the Gerrit
Docker Build file for Gen AI Microservice - This Gerrit contains the Docker build file for generating the image of GenAi microservice. It also contains the python project file for building the microservice inside the docker container. It also contains Docker file(test.Docker) to compile and generate the unit test image of the microservice.

Yaml file to Spawn container - This Gerrit contains the Yaml file to spawn the container of Gen-AI microservice. It also contains yaml file to run the Unit tests.

Unit test and Mock Library - This Gerrit contains the python unit test and a mock library which will be used to mock the response in case of Unit test.
Change-Id: I8bf2c6fb99bf3a080c2241deec49970ec9dcd827